split files: Created log.py and palette.py
log.py includes the Log class.
palette.py include stuff about constructing indexed images and k-means

Issue: #22
